requirements.txt
borax==3.5.0
Pillow==8.4.0
tqdm==4.62.0
